// src/api/types.d.ts
export namespace NQuakeAPI {
// ...
    export interface GAUNTLET {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface MACHINEGUN {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface MACHINEGUN_GRADE1 {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface SHOTGUN {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface SHOTGUN_GRADE1 {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface NAILGUN {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface NAILGUN_GRADE1 {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface ROCKET_LAUNCHER {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface LIGHTNING_GUN {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface RAILGUN {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
    export interface LAGBOLT {
        hits: number;
        shots: number;
        kills: number;
        headshots: number;
        damage: number;
    }
// ...
    export interface IGamesSummaryMatchesItem {
        id: string;
        time: string;
        mapName: string;
        rank: number;
        score: number[];
        gameMode: string;
        won: boolean;
        xp: number;
        kdr: number;
        totalDamage: number;
        weaponAccuracy: IGamesSummaryWeaponAccuracy;
    }
    export interface IGamesSummaryWeaponAccuracy {
        GAUNTLET: number;
        MACHINEGUN: number;
        MACHINEGUN_GRADE1: number;
        SHOTGUN: number;
        SHOTGUN_GRADE1: number;
        NAILGUN: number;
        NAILGUN_GRADE1: number;
        ROCKET_LAUNCHER: number;
        LIGHTNING_GUN: number;
        RAILGUN: number;
        LAGBOLT: number;
    }

    export  interface IGamesSummary {
        matches: IGamesSummaryMatchesItem[]
    }
}